requirements.txt
google-api-python-client==2.143.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
gspread==6.1.4
gspread-formatting==1.2.0
oauth2client==4.1.3
pandas==2.2.2
numpy==1.26.4
